+++ Technologies/Technologies/Screens/TechnologyGridView/TechnologiesGridViewModel.swift
@@ -16,4 +16,9 @@ final class TechnologiesGridViewModel: ObservableObject {
     }
     
     @Published var isShowingDetailView = false
+    
+    let columns: [GridItem] = [GridItem(.flexible()),
+                               GridItem(.flexible()),
+                               GridItem(.flexible())
+    ]
 }
